Major restructuring. Still need to DRY up lib.rs
