+++ src/main.rs
@@ -1,10 +1,17 @@
+use anyhow::Result;
 use clap::Parser;
-use cri_headings::{args::Args, run};
+use cri_headings::args::Args;
+use cri_headings::*;
+use std::fs;
+use std::sync::Arc;
 
-fn main() {
-    let args = Args::parse();
-    if let Err(e) = run(args) {
-        eprintln!("{}", e);
-        std::process::exit(1);
+#[tokio::main]
+pub async fn main() -> Result<()> {
+    let args = Arc::new(Args::parse());
+    fs::create_dir_all(&args.output_dir)?;
+
+    match args.years.len() {
+        1 => blocking_run(args.clone()),
+        _ => async_run(args.clone()).await,
     }
 }
